feat(material-test-generator): revise layer name, move folder into dialogs
