+++ src/web/app/components/dialogs/MaterialTestGeneratorPanel/BlockSetting.ts
