+++ src/web/app/components/dialogs/MaterialTestGeneratorPanel/WorkAreaInfo.module.scss
