+++ src/web/app/components/dialogs/MaterialTestGeneratorPanel/WorkAreaInfo.tsx
