+++ src/web/app/components/dialogs/MaterialTestGeneratorPanel/generateSvgInfo.ts
