+++ src/web/app/components/dialogs/MaterialTestGeneratorPanel/index.module.scss
